Add wasm tests using was-bindgen-test and wasm-pack (#346)
* Add wasm tests using was-bindgen-test and wasm-pack

* Remove custom wasm feature

* Add #[wasm_bindgen_test] to remaining tests

* Try to fix CI

---------

+++ examples/custom_header.rs
@@ -7,7 +7,7 @@ use jsonwebtoken::{decode, encode, Algorithm, DecodingKey, EncodingKey, Header, 
 struct Claims {
     sub: String,
     company: String,
-    exp: usize,
+    exp: u64,
 }
 
 fn main() {
